pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "iblog"
version = "0.2.0"
description = "轻量级静态博客生成工具"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=4.0.0",
    "jinja2>=3.1.6",
    "loguru>=0.7.3",
    "markdown-it-py[linkify,plugins]>=4.0.0",
    "pydantic>=2.12.5",
    "python-frontmatter>=1.1.0",
    "pyyaml>=6.0",
    "typer>=0.23.0",
]

[project.scripts]
iblog = "iblog.cli.build:app"

[tool.hatch.build.targets.wheel]
packages = ["src/iblog"]
